requirements.txt
streamlit>=1.32.0
openai>=1.14.0
PyPDF2>=3.0.1
python-docx>=1.1.0
faiss-cpu>=1.7.4
pandas>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
textstat>=0.7.3
numpy>=1.26.0
